+++ CoinGecko/Entities/Response/Coins/CoinFullData.cs
@@ -61,15 +61,15 @@ namespace CoinGecko.Entities.Response.Coins
     {
         [JsonProperty("roi")] public Roi Roi { get; set; }
 
-        [JsonProperty("current_price")] public Dictionary<string, double?> CurrentPrice { get; set; }
+        [JsonProperty("current_price")] public Dictionary<string, decimal?> CurrentPrice { get; set; }
 
-        [JsonProperty("market_cap")] public Dictionary<string, double?> MarketCap { get; set; }
+        [JsonProperty("market_cap")] public Dictionary<string, decimal?> MarketCap { get; set; }
 
-        [JsonProperty("total_volume")] public Dictionary<string, double?> TotalVolume { get; set; }
+        [JsonProperty("total_volume")] public Dictionary<string, decimal?> TotalVolume { get; set; }
 
-        [JsonProperty("high_24h")] public Dictionary<string, double?> High24H { get; set; }
+        [JsonProperty("high_24h")] public Dictionary<string, decimal?> High24H { get; set; }
 
-        [JsonProperty("low_24h")] public Dictionary<string, double?> Low24H { get; set; }
+        [JsonProperty("low_24h")] public Dictionary<string, decimal?> Low24H { get; set; }
 
         [JsonProperty("price_change_percentage_7d")]
         public string PriceChangePercentage7D { get; set; }
@@ -90,45 +90,45 @@ namespace CoinGecko.Entities.Response.Coins
         public string PriceChangePercentage1Y { get; set; }
 
         [JsonProperty("price_change_24h_in_currency")]
-        public Dictionary<string, double> PriceChange24HInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChange24HInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_1h_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage1HInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage1HInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_24h_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage24HInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage24HInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_7d_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage7DInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage7DInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_14d_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage14DInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage14DInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_30d_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage30DInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage30DInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_60d_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage60DInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage60DInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_200d_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage200DInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage200DInCurrency { get; set; }
 
         [JsonProperty("price_change_percentage_1y_in_currency")]
-        public Dictionary<string, double> PriceChangePercentage1YInCurrency { get; set; }
+        public Dictionary<string, decimal> PriceChangePercentage1YInCurrency { get; set; }
 
         [JsonProperty("market_cap_change_24h_in_currency")]
-        public Dictionary<string, double> MarketCapChange24HInCurrency { get; set; }
+        public Dictionary<string, decimal> MarketCapChange24HInCurrency { get; set; }
 
         [JsonProperty("market_cap_change_percentage_24h_in_currency")]
-        public Dictionary<string, double> MarketCapChangePercentage24HInCurrency { get; set; }
+        public Dictionary<string, decimal> MarketCapChangePercentage24HInCurrency { get; set; }
     }
 
     public class Roi
     {
-        [JsonProperty("times")] public double? Times { get; set; }
+        [JsonProperty("times")] public decimal? Times { get; set; }
 
         [JsonProperty("currency")] public string Currency { get; set; }
 
-        [JsonProperty("percentage")] public double? Percentage { get; set; }
+        [JsonProperty("percentage")] public decimal? Percentage { get; set; }
     }
 }
